requirements.txt
streamlit==1.31.1
jinja2>=3.1.3
weasyprint>=60.2
python-dotenv==1.0.0
langchain>=0.1.12
langchain-groq>=0.1.0
groq>=0.4.2
pyrebase4==4.7.1
setuptools<81
<<<<<<< HEAD
setuptools
=======
setuptools<81.0.0
firebase-admin==6.4.0
requests==2.31.0
gcloud==0.18.3
oauth2client==4.1.3
>>>>>>> 573db29d34981b8efea90231a1d378da070482c5

